requirements.txt
langchain>=0.1.10
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
weaviate-client>=4.9.3
pydantic>=2.5.0
python-dotenv>=1.0.0
pypdf>=3.17.0
python-pptx>=0.6.21
pillow>=10.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3
numpy>=1.26.2
pandas>=2.1.3
scikit-learn>=1.3.2
torch>=2.1.1
transformers>=4.36.2
sentence-transformers>=2.2.2
ollama>=0.1.21
requests>=2.31.0
mlflow>=2.10.0
chromadb>=0.4.22
